Pass default parameters to player constructor
Player cannot work correctly without an event handler, so intead of
registering after the player is instantianted, it is passed as a
constructor parameter.

In the case of the frame renderer, if it non rendered is set, the
video track in the player should be disabled. This patch pass the
frame renderer decorated with an Option, so if it is none, the video
track is disabled, thus removing disable_video() method.

Fixes: #258