Initialize media in the background (#405)
Move initialization of the media engine to a background thread. This
also replaced the complex INITIALIZER / INSTANCE combo with a single
synchronous `OnceCell`. Finally, instead of taking in the `Backend` as
an argument, take in a factory function. This allows the `Backend` to
also be initialized off the main thread.

This allows GStreamer plugin cache initialization to happen off the main
thread and removes unsafe code in initialization.

Co-authored-by: Rakhi Sharma <atbrakhi@igalia.com>