Make ServoMedia::get infallible
It was already infallible before, but returned
a Result type for unknown reasons. The relevant
Code hasn't been touched for 7 years.

I think the original intention was to make backend
creation fallible, but as mentioned before
it currently isn't, and even if it were it
seems more reasonable to return a Result from
ServoMedia::init.

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>