gst: Set the playback rate with validation
The `gst_play_set_rate` function not just setting the playback
rate property in GstPlay but also make new `seek` request with
the current position, `seek done` event afterwards and
changing the player state to `paused`.

Need to add validation of the incoming playback rate to prevent
unnessesary `seek` requests inside the media pipeline.

This unexpected (for the user agent) functionality of
`gst_play_set_rate` is required for the media element
(HTMLMediaElement) and as workaround will be replaced to
`gst_play_seek` which do the same.
https://github.com/servo/servo/issues/40740

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>