gstreamer: Decouple ownership of the GstPlay/AppSrc from signal callbacks
The associated signal callbacks shouldn't capture strong or weak
references of a owner `GstPlay` object with which they are interact
otherwise it leads to leaked underlying media resources.

Note that upgrading the captured weak reference for a `GstPlay` object
on another thread (the `GstPlay` streaming thread) is not safely because
a `GstPlay` object could be destroyed on that thread after signal callback
handling with mutex panic on the `gst_play_finalize`.

To solve the `GstPlay` ownership issue between client and Gstreamer was
introduced the player shared state with minimal required states
(`metadata`, `pending playback rate`, `play state`, ...).

Anyway the captured `GstPlay` weak reference will be used on sync `source`
setup stage which is locking the client thread until a source will be ready.

For `GstAppSrc` and associated `seek_data` callback a capture of a weak reference
will be used with explicit set `empty` callbacks on a player destruction.

Fixes: https://github.com/servo/media/issues/455

Signed-off-by: Andrei Volykhin <andrei.volykhin@gmail.com>